Initialise Readme.md
Hi guys,
Please review and adjust accordingly. Jay and dev team, I am unsure of the setup of the dev project for Covidathon. I am assuming we are using the same dev project structure as that of github\Covid19, that is the 'Getting Started' section of the Read Me. If it is incorrect please update accordingly.

Also, the license is to be added.